backend/ad_platform_crew/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "admob_crew"
version = "1.0.0"
description = "Agentic AdMob Control Plane - Conversational interface for AdMob management"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Koushik Ks"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    "langchain-anthropic>=0.3.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
admob-crew = "ad_platform_crew.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
